+++ yd_database_downloader.py
@@ -11,7 +11,7 @@ Ce module gère l'ensemble du processus :
 
 Interface graphique intégrée dans QGIS avec barre de progression.
 
-Compatible avec plugin v2.0.0 et 2.1.0
+Compatible avec plugin v2.0.0, 2.1.0 et 2.2.0
 """
 
 import sqlite3